requirements.txt
django==5.2.5
djangorestframework
django-environ
dj-database-url
python-decouple
psycopg2-binary
gunicorn
celery
redis
django-celery-beat
sentence-transformers
faiss-cpu  # utilisé pour les embeddings
pgvector